Create workflow computers from structure Add timeout condition to simulator Add MVP editor for structure

+++ structure_optimizer/src/main/java/u_szeged/inf/fog/structure_optimizer/structures/ComputerSpecification.java
@@ -1,4 +1,4 @@
 package u_szeged.inf.fog.structure_optimizer.structures;
 
-public record ComputerSpecification(String name, double cores, double processingPerTick, long memory) {
+public record ComputerSpecification(String name, int cores, double processingPerTick, long memory) {
 }
